Integrate DateDisplay into Card component
